Add common code to FileUtil
