+++ src/main/java/io/github/jeddict/ai/util/FileUtil.java
@@ -18,7 +18,11 @@ package io.github.jeddict.ai.util;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import javax.swing.text.Document;
+import org.netbeans.api.project.Project;
 import org.openide.cookies.EditorCookie;
 import org.openide.cookies.SaveCookie;
 import org.openide.filesystems.FileObject;
@@ -51,8 +55,7 @@ public class FileUtil {
             }
         }
     }
-    
-    
+
     public static FileObject createTempFileObject(String name, String content) throws IOException {
         File tempFile = File.createTempFile("GenAI-" + name, ".java");
         tempFile.deleteOnExit();
@@ -63,7 +66,7 @@ public class FileUtil {
         FileObject fileObject = org.openide.filesystems.FileUtil.toFileObject(tempFile);
         return fileObject;
     }
-    
+
     public static String getLatestContent(FileObject fileObject) {
         try {
             DataObject dataObject = DataObject.find(fileObject);
@@ -80,4 +83,62 @@ public class FileUtil {
             return null;
         }
     }
+
+    public static FileObject getFileObject(Project project, String path) {
+        java.io.File file = resolvePath(project, path).toFile();
+        return org.openide.filesystems.FileUtil.toFileObject(file);
+    }
+    
+    public static String readContent(Project project, String path) throws IOException {
+        return Files.readString(resolvePath(project, path));
+    }
+
+    public static Path resolvePath(Project project, String path) {
+        Path p = Paths.get(path);
+        if (!p.isAbsolute() && project != null) {
+            FileObject projectDir = project.getProjectDirectory();
+            if (projectDir != null) {
+                return Paths.get(org.openide.filesystems.FileUtil.toFile(projectDir).getAbsolutePath(), path)
+                        .toAbsolutePath().normalize();
+            }
+        }
+        return p.toAbsolutePath().normalize();
+    }
+    
+    /**
+     * Utility method that opens a NetBeans document for a given path and
+     * applies the specified action.
+     * <p>
+     * This centralizes error handling and ensures that the document is saved
+     * after the action is applied.
+     *
+     * @param path the relative or absolute path to the file
+     * @param action the action to perform on the opened {@link Document}
+     * @param save save the doc after post action
+     * @return the result of the action, or an error message
+     */
+    public static String withDocument(Project project, String path, DocAction action, boolean save) {
+        try {
+            FileObject fo = getFileObject(project, path);
+            if (fo == null) {
+                return "File not found: " + path;
+            }
+
+            DataObject dobj = DataObject.find(fo);
+            EditorCookie cookie = dobj.getLookup().lookup(EditorCookie.class);
+            if (cookie == null) {
+                return "No editor available for: " + path;
+            }
+
+            Document doc = cookie.openDocument();
+            String result = action.apply(doc);
+            if (save) {
+                cookie.saveDocument();
+            }
+            return result;
+        } catch (Exception e) {
+            return "Operation failed: " + e.getMessage();
+        }
+    }
+
 }
